move knex boostrap to utils folder
